% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECx.R
\name{ECx}
\alias{ECx}
\title{Calculate ECx}
\usage{
ECx(x, y)
}
\arguments{
\item{model}{A model object, typically from \code{glm}, for which the ECx will
be calculated.}
}
\value{
The calculated ECx value.
}
\description{
Description of what ECx calculation does, including any important details
users should know.
}
\examples{
# Here, you could include a simple example of how to use the ECx function.
}
