% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hav4_documnetation.R
\docType{data}
\name{hav4}
\alias{hav4}
\title{hav4}
\format{
A data frame with 48 rows of a continous treatment (raw.x), survivorship/success (suc), and total (tot)
\describe{
\item{column1}{treatment concentration}
\item{column2}{success}
\item{column2}{total}
...
}
}
\source{
Source of your dataset (if applicable).
}
\usage{
hav4
}
\description{
A brief description of my dataset, including its source, what it represents,
and any other relevant information.
}
\keyword{datasets}
